% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overfit.R
\name{overfitLDA}
\alias{overfitLDA}
\title{Run an example of overfitting using LDA}
\usage{
overfitLDA()
}
\value{
None This function starts a shiny example.
}
\description{
Based an example from Di Cook, Module 2 of SSBID 2015.
}

