% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overfitme.R
\docType{package}
\name{overfitme}
\alias{overfitme}
\alias{overfitme-package}
\title{Examples of overfitting}
\description{
Examples of overfitting. Currently only contains an example from
Di Cook of overfitting LDA with simulated data.
}

