% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overfit.R
\name{plotLDAPreds}
\alias{plotLDAPreds}
\title{Plot LDA predictions}
\usage{
plotLDAPreds(tr.preds, tst.preds, xmin, xmax)
}
\arguments{
\item{tr.preds}{Training predictions data.frame}

\item{tst.preds}{Test predictions data.frame}

\item{xmin}{ignored. Future work.}

\item{xmax}{ignored. Future work.}
}
\value{
ggplot object
}
\description{
Plot LDA predictions
}

