% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overfit.R
\name{plotLDAScalings}
\alias{plotLDAScalings}
\title{Plot LDA coefficients}
\usage{
plotLDAScalings(scalings)
}
\arguments{
\item{scalings}{Data.frame of LDA coefficients}
}
\value{
ggplot object
}
\description{
Plot LDA coefficients
}

