% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overfit.R
\name{server}
\alias{server}
\title{Shiny server function}
\usage{
server(input, output)
}
\arguments{
\item{input}{Something passed in by shiny}

\item{output}{Something used by shiny}
}
\value{
None This runs until cows come home
}
\description{
Shiny server function
}

